Update RyRyCryptor to v1.0.1 Refactor and reorganise CarDodgeCore, CarDodgeMain and CarInfo structures and classes with their own respective headers Remove Definitions.h file, use better alternative from ZeeTerminal source Change title to v0.6.0
