+++ CarDodge.vcxproj.filters
@@ -18,6 +18,24 @@
     <ClCompile Include="CarDodge.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="CarDodge_CodeFiles\CarDodgeCore\CarDodgeCore.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="CarDodge_CodeFiles\CarDodgeMain\CarDodgeMain.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="CarDodge_CodeFiles\CarInfo\CarInfo.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\ZeeTerminal\ZeeTerminal\Core\ZTFormattingDefinitions.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="CarDodge_Definitions\ZeeTerminalCore.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="CarDodge_Definitions\ColourConv.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="CarDodge_Definitions\Definitions.h">
@@ -27,6 +45,21 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="RyRyCryptor\RyRyCryptor.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="CarDodge_CodeFiles\CarDodgeCore\CarDodgeCore.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="CarDodge_CodeFiles\CarDodgeMain\CarDodgeMain.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="CarDodge_CodeFiles\CarInfo\CarInfo.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\ZeeTerminal\ZeeTerminal\Core\ZTFormattingDefinitions.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="CarDodge_Definitions\ZeeTerminalCore.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
