+++ CarDodge_Definitions/ZTFormattingDefinitions.cpp
@@ -1,0 +1,75 @@
+#include <string>
+#include "ZTFormattingDefinitions.h"
+
+// Normal colour definitions - ANSI-only
+const std::string BLK_ANSI = "0;0;0";
+const std::string RED_ANSI = "230;0;0";
+const std::string GRN_ANSI = "22;198;12";
+const std::string YLW_ANSI = "231;186;0";
+const std::string BLU_ANSI = "0;0;255";
+const std::string MAG_ANSI = "136;23;152";
+const std::string CYN_ANSI = "58;150;221";
+const std::string WHT_ANSI = "215;215;215"; // Used to be 242;242;242
+// Bright colour definitions - ANSI-only
+const std::string GRAY_ANSI = "118;118;118";
+const std::string LRED_ANSI = "251;96;127"; // 231;72;86 is normal
+const std::string LGRN_ANSI = "0;255;0";
+const std::string LYLW_ANSI = "255;255;0";
+const std::string LBLU_ANSI = "59;120;255";
+const std::string LMAG_ANSI = "180;0;158";
+const std::string LCYN_ANSI = "97;214;214";
+const std::string LWHT_ANSI = "255;255;255";
+
+
+// Normal colour definitions - WIN32
+const std::string BLK_WIN32 = "0";
+const std::string BLU_WIN32 = "1";
+const std::string GRN_WIN32 = "2";
+const std::string CYN_WIN32 = "3";
+const std::string RED_WIN32 = "4";
+const std::string MAG_WIN32 = "5";
+const std::string YLW_WIN32 = "6";
+const std::string WHT_WIN32 = "7";
+// Bright colour definitions - WIN32
+const std::string GRAY_WIN32 = "8";
+const std::string LBLU_WIN32 = "9";
+const std::string LGRN_WIN32 = "10";
+const std::string LCYN_WIN32 = "11";
+const std::string LRED_WIN32 = "12";
+const std::string LMAG_WIN32 = "13";
+const std::string LYLW_WIN32 = "14";
+const std::string LWHT_WIN32 = "15";
+
+
+// Normal colour definitions
+std::string BLK = "0;0;0";
+std::string RED = "230;0;0";
+std::string GRN = "22;198;12";
+std::string YLW = "231;186;0";
+std::string BLU = "0;0;255";
+std::string MAG = "136;23;152";
+std::string CYN = "58;150;221";
+std::string WHT = "215;215;215"; // Used to be 242;242;242
+
+// Bright colour definitions
+std::string GRAY = "118;118;118";
+std::string LRED = "251;96;127"; // 231;72;86 is normal
+std::string LGRN = "0;255;0";
+std::string LYLW = "255;255;0";
+std::string LBLU = "59;120;255";
+std::string LMAG = "180;0;158";
+std::string LCYN = "97;214;214";
+std::string LWHT = "255;255;255";
+
+// Test formatting definitions as a string (easier for output)
+std::string ULINE_STR = "\x1b[4m";
+std::string BOLD_STR = "\x1b[1m";
+std::string BLINK_STR = "\x1b[5m";
+std::string STRIKE_STR = "\x1b[9m";
+
+// Text formatting strings (reset) - as a string (easier for output)
+std::string NOULINE_STR = "\x1b[24m";
+std::string NOBOLD_STR = "\x1b[22m";
+std::string NOBLINK_STR = "\x1b[25m";
+std::string NOSTRIKE_STR = "\x1b[29m";
+
