+++ CarDodge_Definitions/ZTFormattingDefinitions.h
@@ -1,0 +1,75 @@
+#pragma once
+#include <string>
+
+// Normal colour definitions - ANSI-only
+extern const std::string BLK_ANSI;
+extern const std::string RED_ANSI;
+extern const std::string GRN_ANSI;
+extern const std::string YLW_ANSI;
+extern const std::string BLU_ANSI;
+extern const std::string MAG_ANSI;
+extern const std::string CYN_ANSI;
+extern const std::string WHT_ANSI;
+// Bright colour definitions - ANSI-only
+extern const std::string GRAY_ANSI;
+extern const std::string LRED_ANSI;
+extern const std::string LGRN_ANSI;
+extern const std::string LYLW_ANSI;
+extern const std::string LBLU_ANSI;
+extern const std::string LMAG_ANSI;
+extern const std::string LCYN_ANSI;
+extern const std::string LWHT_ANSI;
+
+
+// Normal colour definitions - WIN32
+extern const std::string BLK_WIN32;
+extern const std::string BLU_WIN32;
+extern const std::string GRN_WIN32;
+extern const std::string CYN_WIN32;
+extern const std::string RED_WIN32;
+extern const std::string MAG_WIN32;
+extern const std::string YLW_WIN32;
+extern const std::string WHT_WIN32;
+// Bright colour definitions - WIN32
+extern const std::string GRAY_WIN32;
+extern const std::string LBLU_WIN32;
+extern const std::string LGRN_WIN32;
+extern const std::string LCYN_WIN32;
+extern const std::string LRED_WIN32;
+extern const std::string LMAG_WIN32;
+extern const std::string LYLW_WIN32;
+extern const std::string LWHT_WIN32;
+
+
+// Normal colour definitions
+extern std::string BLK;
+extern std::string RED;
+extern std::string GRN;
+extern std::string YLW;
+extern std::string BLU;
+extern std::string MAG;
+extern std::string CYN;
+extern std::string WHT;
+
+// Bright colour definitions
+extern std::string GRAY;
+extern std::string LRED;
+extern std::string LGRN;
+extern std::string LYLW;
+extern std::string LBLU;
+extern std::string LMAG;
+extern std::string LCYN;
+extern std::string LWHT;
+
+// Test formatting definitions as a string (easier for output)
+extern std::string ULINE_STR;
+extern std::string BOLD_STR;
+extern std::string BLINK_STR;
+extern std::string STRIKE_STR;
+
+// Text formatting strings (reset) - as a string (easier for output)
+extern std::string NOULINE_STR;
+extern std::string NOBOLD_STR;
+extern std::string NOBLINK_STR;
+extern std::string NOSTRIKE_STR;
+
